+++ noodles-csi/src/async/writer.rs
@@ -176,6 +176,7 @@ where
         writer,
         depth,
         reference_sequence.bins(),
+        reference_sequence.binned_index(),
         reference_sequence.metadata(),
     )
     .await
@@ -185,6 +186,7 @@ async fn write_bins<W>(
     writer: &mut W,
     depth: u8,
     bins: &IndexMap<usize, Bin>,
+    index: &IndexMap<usize, bgzf::VirtualPosition>,
     metadata: Option<&Metadata>,
 ) -> io::Result<()>
 where
@@ -203,8 +205,9 @@ where
 
     writer.write_i32_le(n_bin).await?;
 
-    for (&id, bin) in bins {
-        write_bin(writer, id, bin).await?;
+    for (id, bin) in bins {
+        let first_record_start_position = index.get(id).copied().unwrap_or_default();
+        write_bin(writer, *id, first_record_start_position, bin).await?;
     }
 
     if let Some(m) = metadata {
@@ -214,14 +217,19 @@ where
     Ok(())
 }
 
-async fn write_bin<W>(writer: &mut W, id: usize, bin: &Bin) -> io::Result<()>
+async fn write_bin<W>(
+    writer: &mut W,
+    id: usize,
+    first_record_start_position: bgzf::VirtualPosition,
+    bin: &Bin,
+) -> io::Result<()>
 where
     W: AsyncWrite + Unpin,
 {
     let bin_id = u32::try_from(id).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32_le(bin_id).await?;
 
-    let loffset = u64::from(bin.loffset());
+    let loffset = u64::from(first_record_start_position);
     writer.write_u64_le(loffset).await?;
 
     write_chunks(writer, bin.chunks()).await?;