+++ noodles-csi/src/index/reference_sequence/bin/builder.rs
@@ -1,21 +1,14 @@
-use noodles_bgzf as bgzf;
-
 use super::{Bin, Chunk};
 
 /// A CSI index reference sequence bin builder.
-#[derive(Debug)]
+#[derive(Debug, Default)]
 pub struct Builder {
-    pub(crate) loffset: bgzf::VirtualPosition,
     chunks: Vec<Chunk>,
 }
 
 impl Builder {
     /// Adds or merges a chunk.
     pub fn add_chunk(&mut self, chunk: Chunk) {
-        if chunk.start() < self.loffset {
-            self.loffset = chunk.start();
-        }
-
         if let Some(last_chunk) = self.chunks.last_mut() {
             if chunk.start() <= last_chunk.end() {
                 *last_chunk = Chunk::new(last_chunk.start(), chunk.end());
@@ -29,31 +22,16 @@ impl Builder {
     /// Builds a bin.
     pub fn build(self) -> Bin {
         Bin {
-            loffset: self.loffset,
             chunks: self.chunks,
-        }
-    }
-}
-
-impl Default for Builder {
-    fn default() -> Self {
-        Self {
-            loffset: bgzf::VirtualPosition::MAX,
-            chunks: Vec::new(),
         }
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use noodles_bgzf as bgzf;
 
-    #[test]
-    fn test_default() {
-        let builder = Builder::default();
-        assert_eq!(builder.loffset, bgzf::VirtualPosition::MAX);
-        assert!(builder.chunks.is_empty());
-    }
+    use super::*;
 
     #[test]
     fn test_add_chunk() {
@@ -119,7 +97,6 @@ mod tests {
         let actual = builder.build();
 
         let expected = Bin {
-            loffset: bgzf::VirtualPosition::from(5),
             chunks: vec![Chunk::new(
                 bgzf::VirtualPosition::from(5),
                 bgzf::VirtualPosition::from(13),