+++ noodles-tabix/src/async/reader.rs
@@ -188,7 +188,7 @@ where
             metadata.replace(m).is_some()
         } else {
             let chunks = read_chunks(reader).await?;
-            let bin = Bin::new(bgzf::VirtualPosition::default(), chunks);
+            let bin = Bin::new(chunks);
             bins.insert(id, bin).is_some()
         };