+++ noodles-tabix/src/reader.rs
@@ -143,7 +143,7 @@ where
             let chunks =
                 read_chunks(reader).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
-            let bin = Bin::new(bgzf::VirtualPosition::default(), chunks);
+            let bin = Bin::new(chunks);
 
             bins.insert(id, bin).is_some()
         };